components/ai-agent/requirements.txt
langchain
langchain-community
langchain-chroma
langchain-text-splitters
langchain-ollama
langchain-huggingface
sentence-transformers
langchain-groq
tavily-python
chromadb
fastapi
uvicorn[standard]
python-dotenv
openpyxl

